requirements.txt
astroid==1.6.1
colorama==0.3.9
Django==1.11
django-filter==1.1.0
django-webpack-loader==0.5.0
djangorestframework==3.7.7
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
pylint==1.8.2
pytz==2018.3
six==1.11.0
wrapt==1.10.11
